refactor(types): split structure.ts into entities/{relationship,structure-entity-type,structure-entity,structure-entity-link,structure-entity-member-link,structure-entity-association}.ts
