+++ packages/types/src/__tests__/encryption.test.ts
@@ -76,6 +76,7 @@ import type { ChatMessage } from "../entities/message.js";
 import type { Note } from "../entities/note.js";
 import type { PollVote } from "../entities/poll-vote.js";
 import type { Poll } from "../entities/poll.js";
+import type { Relationship } from "../entities/relationship.js";
 import type { Group } from "../groups.js";
 import type {
   BucketId,
@@ -90,7 +91,6 @@ import type {
 } from "../ids.js";
 import type { JournalEntry, WikiPage } from "../journal.js";
 import type { LifecycleEvent } from "../lifecycle.js";
-import type { Relationship } from "../entities/relationship.js";
 import type { TimerConfig } from "../timer.js";
 import type { UnixMillis } from "../timestamps.js";
 import type { Serialize } from "../type-assertions.js";
